requirements/base.txt
Django==2.1.4
django-filter==2.0.0
djangorestframework==3.9.0
django-cors-middleware==1.3.1
mysqlclient==1.3.14
